# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediERP - Medical Store ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"

  # ── EXPENSES ──────────────────────────────────────────────
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale (POS checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale with line items
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}/receipt:
    get:
      operationId: getSaleReceipt
      tags: [sales]
      summary: Get receipt data for a sale
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"

  # ── RECEIPT SETTINGS ──────────────────────────────────────
...
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        contactName: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        totalPurchases: { type: number }
        totalExpenses: { type: number }
        createdAt: { type: string }
    VendorInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        contactName: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        notes: { type: string }
    VendorUpdate:
      type: object
      properties:
        name: { type: string }
        contactName: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        notes: { type: string }

    # ── EXPENSE ──────────────────────────────────────────────
...
    Sale:
      type: object
      required: [id, saleNumber, date, subtotal, discount, total, paymentAccountId, lines, createdAt]
      properties:
        id: { type: integer }
        saleNumber: { type: string }
        date: { type: string, format: date }
        customerName: { type: ["string", "null"] }
        subtotal: { type: number }
        discount: { type: number }
        total: { type: number }
        paymentAccountId: { type: integer }
        paymentAccountName: { type: string }
        journalEntryId: { type: ["integer", "null"] }
        lines:
          type: array
          items: { $ref: "#/components/schemas/SaleLine" }
        createdAt: { type: string }
    SaleLine:
      type: object
      required: [id, productId, productName, unitType, quantity, unitPrice, total]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        unitType:
          type: string
        quantity: { type: number }
        unitPrice: { type: number }
        discount: { type: number }
        total: { type: number }
    SaleInput:
      type: object
      required: [date, paymentAccountId, lines]
      properties:
        date: { type: string, format: date }
        customerName: { type: string }
        discount: { type: number }
        paymentAccountId: { type: integer }
        lines:
          type: array
          items: { $ref: "#/components/schemas/SaleLineInput" }
    SaleLineInput:
      type: object
      required: [productId, unitType, quantity, unitPrice]
      properties:
        productId: { type: integer }
        unitType:
          type: string
        quantity: { type: number }
        unitPrice: { type: number }
        discount: { type: number }
    Receipt:
      type: object
      required: [sale, settings]
      properties:
        sale: { $ref: "#/components/schemas/Sale" }
        settings: { $ref: "#/components/schemas/ReceiptSettings" }

    # ── RECEIPT SETTINGS ─────────────────────────────────────
    ReceiptSettings:
      type: object
      required: [id]
      properties:
        id: { type: integer }
        storeName: { type: ["string", "null"] }
        storeAddress: { type: ["string", "null"] }
        storePhone: { type: ["string", "null"] }
        storeEmail: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        footerText: { type: ["string", "null"] }
        showLogo: { type: boolean }
        showAddress: { type: boolean }
        showPhone: { type: boolean }
        showEmail: { type: boolean }
        showTaxInfo: { type: boolean }
        taxNumber: { type: ["string", "null"] }
        currency: { type: string }
...