# lib/api-zod/src/generated/types/saleLine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediERP - Medical Store ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface SaleLine {
  id: number;
  productId: number;
  productName: string;
  unitType: string;
  quantity: number;
  unitPrice: number;
  discount?: number;
  total: number;
}