# lib/api-zod/src/generated/types/saleLineInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediERP - Medical Store ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface SaleLineInput {
  productId: number;
  unitType: string;
  quantity: number;
  unitPrice: number;
  discount?: number;
}